backend/requirements.txt
mistralai==0.1.6
weaviate-client==4.3.b0
httpx==0.25.2
pandas==2.2.1
Flask==3.0.2
flask-cors==4.0.0
groq
joblib